+++ lib/net/dns/rr/aaaa.rb
@@ -43,7 +43,7 @@ module Net
         private
 
           def subclass_new_from_hash(options)
-            if options.has_key? :address
+            if options.has_key?(:address)
               @address = check_address(options[:address])
             else
               raise ArgumentError, ":address field is mandatory"