+++ lib/net/dns/rr/ptr.rb
@@ -13,19 +13,25 @@ module Net
       #
       class PTR < RR
 
-        # Getter for PTR resource
-        def ptr
+        # Gets the PTR value.
+        #
+        # Returns a String.
+        def ptrdname
           @ptrdname.to_s
         end
-        alias_method :ptrdname, :ptr
+
+        alias_method :ptr, :ptrdname
+
+        # Gets the standardized value for this record,
+        # represented by the value of <tt>ptrdname</tt>.
+        #
+        # Returns a String.
+        def value
+          ptrdname.to_s
+        end
+
 
         private
-
-        def check_ptr(str)
-          IPAddr.new str
-        rescue
-          raise ArgumentError, "PTR section not valid"
-        end
 
         def build_pack
           @ptrdname_pack = pack_name(@ptrdname)
@@ -36,24 +42,20 @@ module Net
           @ptrdname_pack
         end
 
-        def get_inspect
-          "#@ptrdname"
-        end
-
         def subclass_new_from_hash(args)
           if args.has_key?(:ptrdname) or args.has_key?(:ptr)
             @ptrdname = args[:ptrdname]
           else
-            raise ArgumentError, ":ptrdname or :ptr field is mandatory but missing"
+            raise ArgumentError, ":ptrdname or :ptr field is mandatory"
           end
         end
 
         def subclass_new_from_string(str)
-          @ptrdname = check_ptr(str)
+          @ptrdname = check_name(str)
         end
 
-        def subclass_new_from_binary(data,offset)
-          @ptrdname, offset = dn_expand(data,offset)
+        def subclass_new_from_binary(data, offset)
+          @ptrdname, offset = dn_expand(data, offset)
           offset
         end
 
@@ -62,7 +64,18 @@ module Net
           def set_type
             @type = Net::DNS::RR::Types.new("PTR")
           end
-        
+
+          def get_inspect
+            value
+          end
+
+
+          def check_name(input)
+            IPAddr.new(str)
+          rescue
+            raise ArgumentError, "Invalid PTR Section `#{input}'"
+          end
+
       end
 
     end